mysite/requirements.txt
asgiref==3.5.0
django-cors-headers
autopep8==1.6.0
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
joblib==1.1.0
pandas==1.4.1
pycodestyle==2.8.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
